refactor(examples): extract shared testing utilities to dedicated library
Move E2E test utilities from apps/examples/shared/testing to
libs/examples-shared-testing to properly follow Nx project structure
and avoid path alias issues.

Changes:
- Create libs/examples-shared-testing with proper Nx configuration
- Move base fixtures, types, and playwright config to the library
- Extract app port configuration to separate file for Angular compatibility
- Update all example apps to use @ng-forge/examples-shared-testing
- Remove hardcoded localhost URLs in favor of centralized APP_PORTS config
- Add scope:examples and type:testing tags for project boundaries
