+++ apps/examples/bootstrap/playwright.config.ts
@@ -1,3 +1,3 @@
-import { createPlaywrightConfig } from '@examples/shared/testing/playwright-config';
+import { createPlaywrightConfig } from '@ng-forge/examples-shared-testing/playwright-config';
 
 export default createPlaywrightConfig(import.meta.url, 'bootstrap-examples');
