+++ libs/examples-shared-testing/README.md
@@ -1,0 +1,3 @@
+# examples-shared-testing
+
+This library was generated with [Nx](https://nx.dev).
