+++ libs/examples-shared-testing/project.json
@@ -1,0 +1,9 @@
+{
+  "name": "examples-shared-testing",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/examples-shared-testing/src",
+  "projectType": "library",
+  "tags": ["scope:examples", "type:testing"],
+  "// targets": "to see all targets run: nx show project examples-shared-testing --web",
+  "targets": {}
+}
