+++ libs/examples-shared-testing/src/lib/base-fixtures.ts
