+++ tsconfig.base.json
@@ -19,7 +19,6 @@
       "@examples/ionic/*": ["apps/examples/ionic/src/app/examples/*"],
       "@examples/material/*": ["apps/examples/material/src/app/examples/*"],
       "@examples/primeng/*": ["apps/examples/primeng/src/app/examples/*"],
-      "@examples/shared/*": ["apps/examples/shared/*"],
       "@ng-doc/generated": ["dist/ng-doc/docs"],
       "@ng-forge/dynamic-forms": ["packages/dynamic-forms/src/index.ts"],
       "@ng-forge/dynamic-forms/integration": ["packages/dynamic-forms/integration/src/public_api.ts"],
@@ -27,7 +26,9 @@
       "@ng-forge/dynamic-forms-bootstrap": ["packages/dynamic-forms-bootstrap/src/index.ts"],
       "@ng-forge/dynamic-forms-ionic": ["packages/dynamic-forms-ionic/src/index.ts"],
       "@ng-forge/dynamic-forms-material": ["packages/dynamic-forms-material/src/index.ts"],
-      "@ng-forge/dynamic-forms-primeng": ["packages/dynamic-forms-primeng/src/index.ts"]
+      "@ng-forge/dynamic-forms-primeng": ["packages/dynamic-forms-primeng/src/index.ts"],
+      "@ng-forge/examples-shared-testing": ["libs/examples-shared-testing/src/index.ts"],
+      "@ng-forge/examples-shared-testing/playwright-config": ["libs/examples-shared-testing/src/lib/playwright-config.ts"]
     }
   },
   "exclude": ["node_modules", "tmp"]
